+++ yachtproject/src/App.css
@@ -74,150 +74,6 @@ p {
   font-family: "Poppins", serif;
 }
 
-/* hero container for image */
-.outer-div {
-  position: relative;
-  min-height: 50vh;
-  min-width: 50vw;
-  background-size: cover;
-  background-position: center;
-}
-
-.outer-div-m {
-  position: relative;
-  min-height: 70vh;
-  background-size: cover;
-  background-position: center;
-}
-
-/* for card in intro section*/ 
-.flex-item {
-  flex: 1 1 90%; /* Default: 4x1 layout (each item takes 50% width) */
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column; /* Make the flex container a column */
-  justify-content: center;
-  padding: 20px;
-  max-width: 40vw;
-  max-height: 45vh;
-}
-
-.custom-hover-card {
-  transition: transform 0.3s ease, box-shadow 0.3s ease;
-}
-
-.custom-hover-card:hover {
-  transform: translateY(-10px) scale(1.05); /* Move the card up slightly and scale it */
-  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2); /* Add a stronger shadow on hover */
-  border-color: white; /* Change border color to a custom color (tomato in this case) */
-  border-width: 4px;
-}
-
-.flex-container-r {
-  display: flex;
-  flex-direction: column; /* Stack cards vertically */
-  align-items: center; /* Center align cards */
-
-}
-
-.flex-item-r {
-  width: 100%; /* Cards take full width on mobile */
-  max-width: 90%; /* Optional to leave some margin */
-
-  margin-top: 30px;
-}
-
-/* Static image for cards */
-.card-image img {
-  width: 100%;
-  height: 250px;
-  object-fit: cover;
-  border-radius: 4px;
-}
-
-/* Modal gallery */
-.modal-image-gallery {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 300px;
-  overflow: hidden;
-}
-
-.modal-image-scroll {
-  display: flex;
-  scroll-snap-type: x mandatory;
-  overflow-x: scroll;
-  width: 100%;
-  height: 100%;
-  scroll-behavior: smooth;
-}
-
-.modal-image-scroll-item {
-  flex: 0 0 100%;
-  height: 100%;
-  scroll-snap-align: center;
-}
-
-.modal-image-scroll-item img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 4px;
-}
-
-/* Scroll buttons */
-.scroll-button {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  background-color: rgba(0, 0, 0, 0.5);
-  color: white;
-  border: none;
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  cursor: pointer;
-  font-size: 1.5rem;
-  z-index: 10;
-}
-
-.scroll-button.left {
-  left: 10px;
-}
-
-.scroll-button.right {
-  right: 10px;
-}
-
-.scroll-button:hover {
-  background-color: rgba(0, 0, 0, 0.7);
-}
-
-
-/* Card image container to ensure uniform image size */
-.card-image {
-  width: 100%;
-  height: 250px; /* Set a consistent height for all card images */
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.card-image img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover; /* Ensures the image covers the container */
-}
-
-.custom-form .ant-form-item-label > label {
-  color: white !important; /* Set the label text color to white */
-  font-size: 16px !important; /* Set font size */
-  font-weight: bold; /* Optional: Make it bold */
-}
-
 /* Footer Section Container */
 .footer-section {
   display: flex;
@@ -293,127 +149,9 @@ p {
   text-decoration: underline;
 }
 
-.responsive-container {
-  display: flex;
-  flex-direction: column; /* Default to 2 rows (vertical) */
-  background-color: black;
-  color: white; /* Text color to make content visible */
-  padding: 20px; /* Optional: Add some padding */
-  gap: 10px; /* Space between the boxes */
-}
-
-/* Each box styling */
-.box {
-  background-color: gray;
-  padding: 20px;
-  text-align: center;
-  border-radius: 20px;
-  
-}
-
-.outer-div-train {
-  background-size: cover;
-}  
-
-.training-container {
-  height: 95vh;
-}
-
-.form-label .ant-form-item-label > label {
-  color: #000 ;
-}
-
-.outer-div-m2 {
-  position: relative;
-  min-width: 50vw;
-  background-size: cover;
-  background-position: center;
-}
-
-
-/* Media query for larger screens (e.g., desktops) */
-@media (min-width: 768px) {
-  .responsive-container {
-    flex-direction: row; /* Change to 1 row and 2 columns */
-  }
-}
-
-/* mobile Specific Tweaks (vertical device is default)*/ 
-@media (min-width: 420px) {
-     
-  
-}
-
-/* Tablet */
-@media (min-width: 600px) {
-}
 
 /* Desktop */
-@media (min-width: 1024px) {
-  .outer-div {
-    position: relative;
-    min-height: 100vh;
-    min-width: 50vw;
-    background-size: cover;
-    background-position: center;
-  }
-
-  .outer-div-m2 {
-    position: relative;
-    min-width: 50vw;
-    background-size: cover;
-    background-position: center;
-  }
-
-  .outer-div-h{
-    max-height: 70vh;
-  }
-
-  .intro-container {
-    min-height: 60vh;
-  }
-
-  .intro-container-r {
-    max-height: 60vh;
-  }
-
-  .intro-text{
-    margin-left: 120px;
-  }
-  .flex-container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 20px; /* Space between cards */
-    margin: 20px; /* Optional margin around the container */
-  }
-  
-  .flex-container {
-    flex-direction: row; /* Stack cards vertically */
-    align-items: center; /* Center the cards */
-  }
-
-  .flex-item {
-    flex: none; /* Cards take full width in vertical layout */
-    width: 100%;
-    min-width: 30vw;
-    /* min-height: 40vh; */
-  }
-
-  .flex-container-r {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    gap: 30px;
-    
-  }
-  
-  .flex-item-r {
-    flex: 1 1 300px; /* Default width for each card */
-    max-width: 300px; /* Limit the width of cards */
-    box-sizing: border-box; /* Ensure padding and border are included in the width */
-  }
-  
+@media (min-width: 1024px) {  
   .footer-section {
     flex-direction: row; /* Arrange all sections in a row */
     justify-content: space-between; /* Distribute sections evenly across the row */
@@ -428,52 +166,4 @@ p {
   .footer-info {
     margin-right: 0; /* Remove right margin for the last item to keep it aligned properly */
   }
-  .submit-button .ant-form-item-control {
-    margin-left: calc((100% - 400px) / 2); /* Center the button based on your form width */
-  }
-  .intouch {
-    margin-right: 30px;
-  }
-  .box{
-    max-width: 65%;
-  }
-  .outer-div-r {
-    position: relative;
-    height: 70vh !important;   
-    width:  100vw;
-    background-size: cover;
-  }
-  .r-hero {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding-top: 5rem;
-  }
-  .services-container {
-    height: 30vh
-  }
-  .outer-div-train {
-    background-size: cover; 
-  }
-  .training-container {
-    height: 43vh;
-  }
-}
-
-@media (min-width: 2000px) {
-  .intro-container-r {
-    max-height: 40vh;
-  }
-
-  .outer-div-r {
-    height: 40vh !important;   
-
-  }
-
-  .outer-div-m2 {
-    max-height: 50vh !important;
-    position: relative;
-  }
-
 }
