+++ yachtproject/src/components/CaptainsSection.module.css
@@ -1,5 +1,5 @@
 .captainsSection {
-  padding: 6rem 0;
+  padding: 3rem 0 1rem;
   background: linear-gradient(135deg, #f8f9fa 0%, #ffffff 100%);
 }
