+++ yachtproject/src/components/ReviewsSection.module.css
@@ -1,5 +1,5 @@
 .reviewsSection {
-  padding: 6rem 0;
+  padding: 3rem 0 1rem;
   background: #ffffff;
 }
