+++ yachtproject/src/components/yacht/YachtGallery.module.css
@@ -7,15 +7,16 @@
   position: relative;
   border-radius: 16px;
   overflow: hidden;
-  margin-bottom: 1rem;
-  height: 500px;
+  margin-bottom: 0.2rem;
+  height: 74.5vh;
   box-shadow: 0 8px 25px rgba(0, 0, 0, 0.1);
 }
 
 .mainImage {
   width: 100%;
   height: 100%;
-  object-fit: cover;
+  object-fit: contain;
+  object-position: center bottom;
 }
 
 .navButton {
@@ -38,7 +39,6 @@
 
 .navButton:hover {
   background: var(--primary-color, #50a4cbff);
-  color: white;
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
 }
 
@@ -54,7 +54,6 @@
   display: flex;
   gap: 1rem;
   overflow-x: auto;
-  padding: 0.5rem 0;
   scrollbar-width: thin;
   scrollbar-color: var(--primary-color, #50a4cbff) #f1f1f1;
 }
