+++ yachtproject/src/pages/Rental.module.css
@@ -1,6 +1,7 @@
 /* Rental Page Styles */
 .rentalPage {
   background: #ffffff;
+  font-family: "Poppins", sans-serif;
 }
 
 .container {
@@ -26,7 +27,7 @@
   left: 0;
   right: 0;
   bottom: 0;
-  background: linear-gradient(135deg, rgba(44, 62, 80, 0.8) 0%, rgba(80, 164, 203, 0.6) 100%);
+  background: linear-gradient(135deg, rgba(44, 62, 80, 0.8) 0%, rgba(108, 191, 229, 0.2) 100%);
   z-index: 1;
 }
 
@@ -43,7 +44,6 @@
   font-size: 3.5rem;
   font-weight: 700;
   margin-bottom: 1.5rem;
-  font-family: "Poppins", sans-serif;
   text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
 }
 
@@ -69,8 +69,8 @@
   display: block;
   font-size: 2.5rem;
   font-weight: 700;
-  color: var(--secondary-color);
-  font-family: "Poppins", sans-serif;
+  color: #daa520;
+  
 }
 
 .statLabel {
@@ -97,7 +97,7 @@
   font-size: 2.5rem;
   color: var(--text-primary);
   margin-bottom: 1.5rem;
-  font-family: "Poppins", sans-serif;
+  
   font-weight: 600;
 }
 
@@ -155,26 +155,26 @@
 
 /* Features Section */
 .featuresSection {
-  padding: 6rem 0;
-  background: #ffffff;
+  padding: 4rem 0;
+    background-color: var(--primary-color);
 }
 
 .sectionHeader {
   text-align: center;
-  margin-bottom: 4rem;
+  margin-bottom: 2rem;
 }
 
 .sectionHeader h2 {
   font-size: 2.5rem;
   color: var(--text-primary);
   margin-bottom: 1rem;
-  font-family: "Poppins", sans-serif;
+  
   font-weight: 600;
 }
 
 .sectionHeader p {
   font-size: 1.1rem;
-  color: var(--text-secondary);
+  color: var(--text-primary);
 }
 
 .featuresGrid {
@@ -209,7 +209,7 @@
   font-size: 1.3rem;
   color: var(--text-primary);
   margin-bottom: 1rem;
-  font-family: "Poppins", sans-serif;
+  
   font-weight: 600;
 }
 
@@ -222,7 +222,7 @@
 /* Packages Section */
 .packagesSection {
   position: relative;
-  padding: 6rem 0;
+  padding: 3rem 0 1rem;
   background-size: cover;
   background-position: center;
 }
@@ -257,7 +257,7 @@
 .ctaContent h2 {
   font-size: 2.5rem;
   margin-bottom: 1.5rem;
-  font-family: "Poppins", sans-serif;
+  
   font-weight: 600;
   text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
 }
