+++ yachtproject/src/pages/Services.module.css
@@ -159,7 +159,7 @@
 /* Services Section */
 .servicesSection {
   padding: 6rem 0;
-  background: #ffffff;
+  background: var(--primary-color);
 }
 
 .sectionHeader {
