+++ yachtproject/src/pages/yachts/YachtPage.module.css
@@ -3,7 +3,7 @@
 .yachtPage {
   max-width: 85vw;
   margin: 0 auto;
-  padding: 2rem 0rem;
+  padding: 1rem 0rem;
   background: white;
   font-family: "Poppins", sans-serif;
 }
